useUser hook added and user component refactored to use the hook for logic
